fix: Norm code
